rename smart Component. v3.
